% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.bam.disc.R
\name{bin.bam.disc}
\alias{bin.bam.disc}
\title{Get discordant read counts from BAM file}
\usage{
bin.bam.disc(bam.file, bin.df, outfile.prefix = NULL,
  appendIndex.outfile = TRUE, min.weight = 0.1, weight.dist = 50,
  chunk.size = 10000, check.chr.name = TRUE, no.checks = FALSE,
  bai.file = NULL)
}
\arguments{
\item{bam.file}{the BAM file}

\item{bin.df}{a data.frame with the information about the bins. Columns 'chr', 'start'
and 'end' are required.}

\item{outfile.prefix}{the prefix of the name of the output file. The suffix '.bgz' will
be appended to this name prefix after compression. In practice this is present in the output of 'init.filenames' function.}

\item{appendIndex.outfile}{if TRUE (default), the results will be appended regularly on
the output file which will be ultimately compressed and indexed. This is recommend when a large number
of bins are analyzed. If FALSE, a data.frame with the bin counts will be returned and no
file is created.}

\item{min.weight}{the minimum weight (when a discordant read is isolated). Default is 0.1.}

\item{weight.dist}{the distance from its neighbor where a read is considered isolated in the weighting process. Default is 50 (bp).}

\item{chunk.size}{the number of bins to analyze at a time (for memory optimization).
Default is 10 000. Reduce this number if memory problems arise.}

\item{check.chr.name}{if TRUE (default), the function will try to check that the
definition of chromosome (e.g. '1' vs 'chr1') are consistent between the bin
definition and the BAM file. If FALSE, the analysis will continue either way.}

\item{no.checks}{if TRUE, won't stop if there are no reads or the chromosome names are inconsistent. Default is FALSE.}

\item{bai.file}{the index file of the BAM. If NULL it will be guessed.}
}
\value{
a list:
\item{bc}{the final output file name if 'appendIndex.outfile' was TRUE; a data.frame with
the bin counts if not.}
\item{nb.reads}{the number of reads counted.}
}
\description{
Count the number of discordant reads from a BAM file in specified bins. By default the output bin counts are writtent into a file, which is then automatically compressed and indexed.
}
\details{
By default, the function tries to check that the bin definition and the BAM file have compatible chromosome names. For example, 'chr1' in both or '1' in both. Moreover it will stop if no reads are found in the BAM file. To switch off these checks use 'no.checks=TRUE'.
}
\author{
Jean Monlong
}

