% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bin.bam.R
\name{bin.bam}
\alias{bin.bam}
\title{Get read counts from BAM file}
\usage{
bin.bam(bam.file, bin.df, outfile.prefix = NULL, appendIndex.outfile = TRUE,
  proper = TRUE, map.quality = 30, chunk.size = 10000,
  check.chr.name = TRUE)
}
\arguments{
\item{bam.file}{the BAM file}

\item{bin.df}{a data.frame with the information about the bins. Columns 'chr', 'start'
and 'end' are required.}

\item{outfile.prefix}{the prefix of the name of the output file. The suffix '.bgz' will
be appended to this name prefix after compression.}

\item{appendIndex.outfile}{if TRUE (default), the results will be appended regularly on
the output file which will be ultimately indexed. This is recommend when a large number
of bins are analyzed. If FALSE, a data.frame with the bin counts will be returned and no
file are created.}

\item{proper}{if TRUE (default), reads with properly mapped pairs are counted. If
FALSE, reads with improper mapping are counted.}

\item{map.quality}{the minimum mapping quality (PHRED) for the reads to count. Default
is 30.}

\item{chunk.size}{the number of bins to analyze at a time (for memory optimization).
Default is 10 000. Reduce this number if memory problems arise.}

\item{check.chr.name}{if TRUE (default), the function will try to check that the
definition of chromosome (e.g. '1' vs 'chr1') are consistent between the bin
definition and the BAM file. If FALSE, the analysis will continue either way.}
}
\value{
a list:
\item{bc}{the final output file name if 'appendIndex.outfile' was TRUE; a data.frame with
the bin counts if not.}
\item{nb.reads}{the number of reads counted.}
}
\description{
Count the number of reads from a BAM file in specified bins.
}
\author{
Jean Monlong
}

