% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/breakpoint.finder.interactive.R
\name{breakpoint.finder.interactive}
\alias{breakpoint.finder.interactive}
\title{Interactive breakpoint finder}
\usage{
breakpoint.finder.interactive(chr, start, end, test.sample, files.df,
  ref.samples, proper = TRUE, nb.cores = 1, bp.res = 1, flanks = 2000,
  related.samples = NULL)
}
\arguments{
\item{chr}{the chromosome}

\item{start}{the start position}

\item{end}{the end position}

\item{test.sample}{the sample with abnormal coverage}

\item{files.df}{a data.frame with the path to the bam files}

\item{ref.samples}{other samples to visualize}

\item{proper}{should proper mapping be counted.}

\item{nb.cores}{number of cores}

\item{bp.res}{default resolution (bp)}

\item{flanks}{default flank size (bp)}

\item{related.samples}{samples related to the `test.sample` (e.g. parents) to be drawn in a special color.}
}
\value{
a list with
\item{bk.df}{a data.frame with the breakpoint coordinates}
\item{graph}{the ggplot object with the final graph}
}
\description{
Interactive tool to fine tune breakpoint location
}
\author{
Jean Monlong
}

