% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/breakpoint.finder.R
\name{breakpoint.finder}
\alias{breakpoint.finder}
\title{Breakpoint finder}
\usage{
breakpoint.finder(bkpt.gr, files.df, test.sample, ref.samples,
  flank.size = 15000, proper = TRUE, map.quality = 30, bp.res = 5,
  slidW.size = 200, slidW.step = 1)
}
\arguments{
\item{bkpt.gr}{a GRanges object with on or several abnormal regions in a sample.}

\item{files.df}{a data.frame with information about the BAM file path for each sample (columns 'sample' and 'bam').}

\item{test.sample}{the name of the sample of interest.}

\item{ref.samples}{the names of the samples to use as reference for the breakpoint detection.}

\item{flank.size}{the size of the flanking region investigated on each side of a call. 1-2 times the bin size is usually appropriate.}

\item{proper}{Should proper mapping be counted. Default is TRUE.}

\item{map.quality}{the minimum mapping quality for the counted reads. Default is 30.}

\item{bp.res}{the base-pair resolution of the estimation. Default is 5bp.}

\item{slidW.size}{the size of the sliding window. Default is 200.}

\item{slidW.step}{the step used when sliding the window. Default is 1.}
}
\value{
the original GRanges object with updated breakpoint location.
}
\description{
Find breakpoints as the location with the most dramatic local coverage differences between a samples and a set of reference samples.
}
\author{
Jean Monlong
}

