% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrplot.R
\name{chrplot}
\alias{chrplot}
\title{Chromosom plot of CNVs}
\usage{
chrplot(cnv.df, type = c("sample", "stacked"), bin.size = 5e+05,
  chr.df = NULL, showSampleNames = FALSE, chr.prefix = FALSE,
  show.all.chr = FALSE)
}
\arguments{
\item{cnv.df}{a data.frame with the CNVs (columns 'chr', 'start', 'end') for each sample (column 'sample')}

\item{type}{the type of graph to display. 'stacked' (default) shows the frequency line-graph across the chromosome. 'sample' shows a heatmap with each row representing one sample.}

\item{bin.size}{the bin size. Default is 1e6.}

\item{chr.df}{a data.frame with the chromosome boundaries.}

\item{showSampleNames}{should the sample names be displayed. Default is FALSE (it quickly becomes unreadable).}

\item{chr.prefix}{Is there a 'chr' prefix in the chromosome names. Default is FALSE}

\item{show.all.chr}{Should all the chromosomes be displayed (even if no CNVs) ? Default is FALSE, i.e. only chr with CNVs are displayed.}
}
\value{
a ggplot graph object
}
\description{
Display a chromosome view of the CNVs across samples
}
\author{
Jean Monlong
}
