% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cn.plot.R
\name{cn.plot}
\alias{cn.plot}
\title{Plot a region copy-number estimates}
\usage{
cn.plot(chr, start, end, files.df, samples = NULL, flanks = 10000,
  highlight.region = TRUE)
}
\arguments{
\item{chr}{the chromosome}

\item{start}{the start position of the region}

\item{end}{the end position of the region}

\item{files.df}{a data.frame with the path to the different files associated with each sample. If 'sample' is absent from 'bc.f' the correct file from 'files.df' will be used. Default is NULL.}

\item{samples}{the set of samples to represent.}

\item{flanks}{the size of the flanking region. Default is 10000.}

\item{highlight.region}{Should the input region highlighted. Default is TRUE.}
}
\value{
a ggplot object
}
\description{
Plot the copy number estimates in a genomic regions across samples.
}
\author{
Jean Monlong
}

