% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cnv.plot.R
\name{cnv.plot}
\alias{cnv.plot}
\title{Plot a region coverage}
\usage{
cnv.plot(chr, start, end, bc.f, norm.stats.f = NULL, sv.df = NULL,
  ref.samples = NULL, boxplot = FALSE, samples = NULL)
}
\arguments{
\item{chr}{the chromosome}

\item{start}{the start position of the region}

\item{end}{the end position of the region}

\item{bc.f}{the path to the bin count file.}

\item{norm.stats.f}{the path to the normalization statistics file.}

\item{sv.df}{the data.frame with the calls.}

\item{ref.samples}{the names of the reference samples.}

\item{boxplot}{should the reference be represented as boxplots. If FALSE, violin plots will be used.}

\item{samples}{the set of samples to represent.}
}
\value{
a ggplot object
}
\description{
Plot the coverage in a genomic regions across samples. The files ('bc.f', 'norm.stats.f') must be ordered and indexed.
}
\author{
Jean Monlong
}

