% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/comp.index.files.R
\name{comp.index.files}
\alias{comp.index.files}
\title{Compress and index BED-like files}
\usage{
comp.index.files(files, outprefix = files, rm.input = TRUE,
  overwrite.out = TRUE)
}
\arguments{
\item{files}{the names of the files to compress and index.}

\item{outprefix}{the prefix to use to name the output files. By default, the original file names.}

\item{rm.input}{Should the original input file be removed after compression/indexing ? Default is TRUE.}

\item{overwrite.out}{Should a compressed file be overwritten if it already exists ? Default is TRUE.}
}
\value{
a character vector confirming the creation of the new files.
}
\description{
Compress and index BED-like files. The files need to be ordered by genomic position and the first three columns must define chromosome location, start and end position.
}
\author{
Jean Monlong
}

