% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.plot.raw.R
\name{coverage.plot.raw}
\alias{coverage.plot.raw}
\title{Plot the raw coverage in a region}
\usage{
coverage.plot.raw(chr, start, end, files.df, samples, proper = TRUE,
  map.quality = 30, bp.res = 1, flanks = NULL)
}
\arguments{
\item{chr}{chromosome}

\item{start}{genomic coordinate}

\item{end}{genomic coordinate}

\item{files.df}{a data.frame with the path to the bam files}

\item{samples}{the samples to represent}

\item{proper}{should proper mapping be counted. Default is TRUE.}

\item{map.quality}{the minimum mapping quality. Default is 30.}

\item{bp.res}{the resolution, in bp. Default is 1 bp.}

\item{flanks}{the amount of bp to add in the flanks.}
}
\value{
a ggplot object
}
\description{
Plot the raw read coverage in a genomic regions across samples.
}
\author{
Jean Monlong
}

