% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coverage.plot.R
\name{coverage.plot}
\alias{coverage.plot}
\title{Plot a region coverage}
\usage{
coverage.plot(chr, start, end, bc.f, norm.stats.f = NULL, sv.df = NULL,
  ref.samples = NULL, boxplot = FALSE, samples = NULL, anno.df = NULL,
  anno.col = "geneName")
}
\arguments{
\item{chr}{the chromosome}

\item{start}{the start position of the region}

\item{end}{the end position of the region}

\item{bc.f}{the path to the bin count file.}

\item{norm.stats.f}{the path to the normalization statistics file.}

\item{sv.df}{the data.frame with the calls.}

\item{ref.samples}{the names of the reference samples.}

\item{boxplot}{should the reference be represented as boxplots. If FALSE, violin plots will be used.}

\item{samples}{the set of samples to represent.}

\item{anno.df}{a data.frame with additional information (e.g. gene annotation) to be added to the graph.}

\item{anno.col}{the name of 'anno.df' column to use to differentiate elements in the graph. E.g. 'geneName' to color the genes in the graph.}
}
\value{
a ggplot object
}
\description{
Plot the coverage in a genomic regions across samples. The files ('bc.f', 'norm.stats.f') must be ordered and indexed.
}
\author{
Jean Monlong
}

