% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createBinBamRscripts}
\alias{createBinBamRscripts}
\title{Create Rscript files for manual bin counting}
\usage{
createBinBamRscripts(files.df, bin.df, dest.folder = ".", proper = TRUE,
  map.quality = 30)
}
\arguments{
\item{files.df}{a data.frame with the information about the files to
use. Columns 'sample', 'bam' and 'bc' are required and should be
present  after running 'initFileNames' function.}

\item{bin.df}{a data.frame with the information about the bins. Columns
'chr', 'start' and 'end' are required.}

\item{dest.folder}{the folder where the R scripts files and other temporary
files are created.}

\item{proper}{if TRUE (default), reads with properly mapped pairs are
counted. If FALSE, reads with improper mapping are counted.}

\item{map.quality}{the minimum mapping quality (PHRED) for the reads
to count. Default is 30.}
}
\value{
an updated data.frame with the information about the files.
Specifically a new column 'binBamRscripts' holds the name of the script
to run for each sample.
}
\description{
Create Rscript file to run manually the bin counting for each sample.
If many samples and/or high coverage, and \code{BatchJobs} package is
not set up, this function quickly creates R scripts than can then be
send manually to a cluster.
}
\author{
Jean Monlong
}

