% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.controls.dist.R
\name{draw.controls.dist}
\alias{draw.controls.dist}
\title{Select control genomic regions for enrichment analysis}
\usage{
draw.controls.dist(cnv.gr, feat.grl, nb.class = 20, nb.cores = 3,
  redo.duplicates = TRUE, seed.nb.max = 1e+05, min.nb.gr = NULL,
  chr.prefix = "", dist.gr = NULL)
}
\arguments{
\item{cnv.gr}{the input regions}

\item{feat.grl}{a list of the GRanges defining the features to fit. E.g. gene annotation, centromere, etc}

\item{nb.class}{the number of size class to speed up the computation. Default is 20.}

\item{nb.cores}{the number of cores to use. Default is 3.}

\item{redo.duplicates}{should duplicate regions be reselected. Default is TRUE.}

\item{seed.nb.max}{the maximum number of genomic position to use as seeds. Default is 1e5.}

\item{min.nb.gr}{the minimum number of control regions. If NULL (default), as many controls as input are simulated.}

\item{chr.prefix}{the chromosome name prefix. Default is "" (no prefix). Other value could be "chr" if chromosome are defined as 'chr1', 'chr2', etc.}

\item{dist.gr}{a GRanges defining the feature for which we want to control the distance to. Default is NULL, i.e. no control.}
}
\value{
a GRanges object defining the control regions
}
\description{
Find random genomic regions with the same properties as a set of input regions. The properties are defined using annotation tracks. For example, we can simulate random regions that overlap genes as much as the input regions. It can be used to perform enrichment analysis while controling for specific properties.
}
\author{
Jean Monlong
}

