% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fdrtool.quantile.R
\name{fdrtool.quantile}
\alias{fdrtool.quantile}
\title{P-values estimation}
\usage{
fdrtool.quantile(z, quant.int = seq(0.4, 1, 0.02), ref.dist.weight = NULL,
  plot = TRUE, z.th = NULL)
}
\arguments{
\item{z}{a vector with the Z-scores}

\item{quant.int}{a vector with the quantiles to test for the estimation of the
null normal variance.}

\item{ref.dist.weight}{the weight (value between 0 and 1) based on the distance to the reference samples.}

\item{plot}{should some graphs be displayed. Default if FALSE.}

\item{z.th}{the Z-score threshold to use to estimate the null normal variance, if 'quant.int=NULL'. If one value the same threshold is used for both duplication and deletion; if a vector, the threshold for duplication / deletions.}
}
\value{
a list with
\item{pval}{the vector of P-values}
\item{qval}{the vector of Q-values / FDR estimates}
\item{sigma.est.dup}{the estimated null distribution variance for positive Z-scores}
\item{sigma.est.del}{the estimated null distribution variance for negative Z-scores}
}
\description{
Compute P-values and Q-values from the Z-score distribution.
}
\author{
Jean Monlong
}
\keyword{internal}

