% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.noncovered.bins.R
\name{filter.noncovered.bins}
\alias{filter.noncovered.bins}
\title{Filter non-covered bins}
\usage{
filter.noncovered.bins(bins.df, files.df, nb.samples = 10,
  bc.med.min = NULL, plot = TRUE)
}
\arguments{
\item{bins.df}{the data.frame with the bins information that will be subsetted.}

\item{files.df}{the data.frame with the file information, e.g. location of the bin count files.}

\item{nb.samples}{the number of samples to use. Default is 10.}

\item{bc.med.min}{If non-NULL, the minimum median coverage to use for the filtering. If NULL (default), this value will be estimated.}

\item{plot}{Display a graph ? Default is TRUE.}
}
\value{
a subset of the input 'bins.df' data.frame
}
\description{
Find bins that are not covered and remove them from the input 'bins.df'. This function is useful for targeted sequencing when the binning was unaware of the targeted regions. In practice, the bin counts across a few samples are used to assess if a bin is covered or not.
}
\details{
If 'bc.med.min' is NULL, the function will quickly model the coverage due to off-target reads and set the minimum threshold as (median+3sd) of the off-target coverage. The off-tagert coverage is approximated by the maximum non-null coverage in bins with median coverage of 0. If a bin has 0 read in more than half of the samples, it's likely a non-covered bins. However sometimes a few samples have some reads there. The distribution of these noise-coverage is used to fix the minimum threshold in median coverage.
}
\author{
Jean Monlong
}
