% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.comparable.bins.R
\name{find.comparable.bins}
\alias{find.comparable.bins}
\title{Find comparable bins between two sample cohort.}
\usage{
find.comparable.bins(msd1, msd2, plot = TRUE)
}
\arguments{
\item{msd1}{a data.frame or path to the file with the normalization stats from the first run.}

\item{msd2}{a data.frame or path to the file with the normalization stats from the second run.}

\item{plot}{should a graph be displayed. Default is TRUE.}
}
\value{
a data.frame with the bins to use for the comparison.
}
\description{
When two cohort have been run separately but the variants have to be compared, this function finds the bins in which both runs had comparable detection power. Or at least it removes most of the bins that are clearly not covered similarly in both runs. H
}
\details{
If 'plot=TRUE' (Default), a graph is displayed. It compares the average coverage in both cohort and shows the thresholds used to filter out bins with different coverage.
}
\author{
Jean Monlong
}

