% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragment.genome.hg19.R
\name{fragment.genome.hg19}
\alias{fragment.genome.hg19}
\title{Fragment hg19 genome}
\usage{
fragment.genome.hg19(bin.size = 1000, slid.window = bin.size,
  chr.prefix = FALSE, XY.chr = FALSE, quiet = FALSE)
}
\arguments{
\item{bin.size}{the size of the bins}

\item{slid.window}{the size of the sliding window. Default is the size of the bin, i.e. not overlapping windows.}

\item{chr.prefix}{should chromosome name be in the form "chr1" instead of "1". Default is FALSE.}

\item{XY.chr}{should chromosome X and Y be included. Default is FALSE. If TRUE, male and female should be analyzed separately, at least for chromosomes X and Y.}

\item{quiet}{should any verbose display be avoided ? Default is FALSE.}
}
\value{
a data.frame with columns 'chr', 'start' and 'end'.
}
\description{
Fragment hg19 genome into consecutive bins of equal size. This function
require package \code{BSgenome.Hsapiens.UCSC.hg19} to be installed.
}
\author{
Jean Monlong
}
