% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fragment.genome.hg19.R
\name{fragment.genome.hp19}
\alias{fragment.genome.hp19}
\title{Fragment hg19 genome}
\usage{
fragment.genome.hp19(bin.size = 1000, chr.prefix = FALSE, XY.chr = FALSE)
}
\arguments{
\item{bin.size}{the size of the bins}

\item{chr.prefix}{should chromosome name be in the form "chr1" instead of "1". Default is FALSE.}

\item{XY.chr}{should chromosome X and Y be included. Default is FALSE. If TRUE, male and female should be analyzed separately.}
}
\value{
a data.frame with columns 'chr', 'start' and 'end'.
}
\description{
Fragment hg19 genome into consecutive bins of equal size. This function
require package \code{BSgenome.Hsapiens.UCSC.hg19} to be installed.
}
\author{
Jean Monlong
}

