% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender.predict.R
\name{gender.predict}
\alias{gender.predict}
\title{Predict the gender of sample(s)}
\usage{
gender.predict(bc.df)
}
\arguments{
\item{bc.df}{a data.frame with bin coverage. Columns must be "chr", "start", "end". If only one sample is inputed the fourth and last column must be "bc"; if multiple samples are inputed each additionnal column represent one sample.}
}
\value{
a vector or data.frame with the gender for the sample(s)
}
\description{
Predict the gender of a sample or a group of samples
}
\author{
Jean Monlong
}
