% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getGC.hg19}
\alias{getGC.hg19}
\title{GC content computation for specific bins}
\usage{
getGC.hg19(bin.df)
}
\arguments{
\item{bin.df}{a data.frame with bin information. Columns 'chr', 'start' and 'end'
are required.}
}
\value{
the same input data.frame with an extra column 'GCcontent'.
}
\description{
Add a column with GC content to a data.frame with 'chr', 'start' and 'end' column.
Requires BSgenome for Hsapiens hg19; if not installed instructions to install it will
be displayed.
}
\author{
Jean Monlong
}

