% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGC.R
\name{getGC}
\alias{getGC}
\title{GC content computation for specific bins}
\usage{
getGC(bins.df,
  genome = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19)
}
\arguments{
\item{bins.df}{a data.frame with bin information. Columns 'chr', 'start' and 'end'
are required.}

\item{genome}{the BSgenome object with the genome. Usually BSgenome.Hsapiens.UCSC.hg19 or BSgenome.Hsapiens.UCSC.hg38.}
}
\value{
the same input data.frame with an extra column 'GCcontent'.
}
\description{
Add a column with GC content to a data.frame with 'chr', 'start' and 'end' column.
Requires BSgenome for Hsapiens hg19; if not instructions to install it will
be displayed.
}
\author{
Jean Monlong
}
