% Generated by roxygen2 (4.0.1): do not edit by hand
\name{initFileNames}
\alias{initFileNames}
\title{Init file names for analysis}
\usage{
initFileNames(files.df, dest.folder = ".", sample.folder = TRUE,
  code = NULL, dest.folder.relative.path = TRUE)
}
\arguments{
\item{files.df}{a data.frame with the information about the files to
use. Columns 'sample' and 'bam' are required.}

\item{dest.folder}{the folder where the files with the bin counts are to
be created.}

\item{sample.folder}{if TRUE (default), a folder is created for each sample
on the specified destination folder.}

\item{code}{if not NULL, a name for the analysis. Useful PopSV is to be run
several times with different parameters. Default is NULL.}

\item{dest.folder.relative.path}{Is the specified folder defined relatively to
the working directory. Default is TRUE. Set to FALSE if providing absolute
paths (recommended).}
}
\value{
an updated data.frame with the information about the files.
Specifically a new column 'bc' and 'bc.gz' holds the name of the bin counts and
compressed bin counts files to be created later.
}
\description{
Init the names for the files to use in the analysis to come. Specifically the
name of the files with the bin counts and compressed bin counts are created.
}
\author{
Jean Monlong
}

