% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localMax.R
\name{localMax}
\alias{localMax}
\title{Local maximum}
\usage{
localMax(x, min.max.prop = 0.1)
}
\arguments{
\item{x}{a vector with numeric values}

\item{min.max.prop}{the minimum height of the maximum to be considered. Default is 0.1, i.e. all maximum at least 10\% as height as the highest maximum.}
}
\value{
a list with the local maximum position and height.
}
\description{
Finds the local maximum.
}
\author{
Jean Monlong
}
\keyword{internal}
