% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medvar.norm.internal.R
\name{medvar.norm.internal}
\alias{medvar.norm.internal}
\title{Median-variance normalization of bin counts}
\usage{
medvar.norm.internal(bc)
}
\arguments{
\item{bc}{a matrix or data.frame with the bin counts (bin x sample).}
}
\value{
a matrix with the normalized bin counts (bin x sample).
}
\description{
The median and deviation to the median is normalized across samples.
}
\author{
Jean Monlong
}
\keyword{internal}
