% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medvar.norm.R
\name{medvar.norm}
\alias{medvar.norm}
\title{Median-variance normalization of bin counts}
\usage{
medvar.norm(bc, ref.samples, bc.support = NULL, z.poisson = FALSE)
}
\arguments{
\item{bc}{a matrix or data.frame with the bin counts (bin x sample).}

\item{ref.samples}{a vector with the names of the samples to be used as reference.}

\item{bc.support}{TODO}

\item{z.poisson}{Should the Z-score be computed as an normal-poisson hybrid (see
Details). Default is FALSE.}
}
\value{
a list with
\item{norm.stats}{a data.frame with some metrics about the normalization of each
bin (row) : coverage average and standard deviation; number of outlier samples}
\item{bc.norm}{a data.frame, similar to the input 'bc.df', with the normalized bin counts.}
}
\description{
The median and deviation to the median is normalized across samples.
}
\author{
Jean Monlong
}
