% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/merge.samples.R
\name{merge.samples}
\alias{merge.samples}
\title{Merge data.frame across samples}
\usage{
\method{merge}{samples}(files.df, samples, files.col, nb.cores = 1)
}
\arguments{
\item{files.df}{a data.frame with information about each output file.}

\item{samples}{the names of the samples to use.}

\item{files.col}{the name of 'files.df' to use for each 'res' element.}

\item{nb.cores}{the number of cores to use.}
}
\value{
the merged data.frame
}
\description{
Merge files for different samples.
}
\author{
Jean Monlong
}

