% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mergeConsBin.cbs.R
\name{mergeConsBin.cbs}
\alias{mergeConsBin.cbs}
\title{Segmentation using CBS}
\usage{
mergeConsBin.cbs(df, pv.th = 0.01)
}
\arguments{
\item{df}{a data.frame with at least 'chr', 'start' and 'end' columns.}

\item{pv.th}{the P-value threshold used for the segmentation and the segment filtering. Default is 0.01.}
}
\value{
a data.frame, similar to the input but with merged rows.
}
\description{
Circular Binary Segmentation of the Z-scores.
}
\author{
Jean Monlong
}

