% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mergeConsBin.simple}
\alias{mergeConsBin.simple}
\title{Simple merge of abnormal bins}
\usage{
mergeConsBin.simple(res.df, col.mean = c("z", "pv", "qv", "fc"))
}
\arguments{
\item{res.df}{a data.frame with the Z-scores. Columns 'chr', 'start', 'end' and 'z'
are required.}

\item{col.mean}{a list of potential columns for which the mean value should be kept
when merging bins.}
}
\value{
a data.frame similar to the input 'res.df' but with an extra 'nbBinCons'
column (the number of bin merged for each event).
}
\description{
Merge consecutive bins (end1+1->start2).
}
\author{
Jean Monlong
}
\keyword{internal}

