% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeConsBin.z.R
\name{mergeConsBin.z}
\alias{mergeConsBin.z}
\title{Merge abnormal consecutive bins}
\usage{
mergeConsBin.z(res.df, fdr.th = 0.05, sd.null = 1, nb.sim = 1e+06,
  stitch.dist = 10000)
}
\arguments{
\item{res.df}{a data.frame with the Z-scores. Columns 'chr', 'start', 'end' and 'z'
are required.}

\item{fdr.th}{the False Discovery Rate threshold.}

\item{sd.null}{the estimated standard deviation of the Z-score null distribution.
Usually, computed during P-value estimation by 'fdrtool.quantile'.}

\item{nb.sim}{the number of simulated Z-scores for the P-value computation.}

\item{stitch.dist}{the stitching distance, i.e. the maximum distance at which two bins will be merged.}
}
\value{
a data.frame similar to the input 'res.df' but with an extra 'nb.bin.cons'
column (the number of bin merged for each event).
}
\description{
Merge consecutive abnormal bins into one larger event. Two consecutive bins
will be merged if the maximum (minimum) of their two Z-scores if significantly
lower (higher) than what is observed from a null distribution. The null distribution
is simulated by taking the maximum (minimum) values of two Normal values. The variance
of the simulated Normal is given as a parameter of the function.
}
\author{
Jean Monlong
}
\keyword{internal}

