% Generated by roxygen2 (4.0.2): do not edit by hand
\name{norm.bc}
\alias{norm.bc}
\title{Normalize bin counts}
\usage{
norm.bc(bc = NULL, files.df = NULL, bins.df = NULL, method = c("tn",
  "pca", "medvar"), ...)
}
\arguments{
\item{bc}{a data.frame with the bin counts.}

\item{files.df}{a data.frame with the names of the files with the the bin counts. Used if 'bc' is NULL.}

\item{bins.df}{a data.frame with the information about the bins. Columns 'chr', 'start'
and 'end' are required.}

\item{method}{the normalization method: 'tn' for targeted normalization; 'pca' for
principal component regression; 'medvar' for median/variance global normalization.}

\item{...}{extra parameters for the normalization methods.}
}
\value{
the output from 'tn.norm'/'pca.norm'/'medvar' function, depending of
the choice of 'method'.
}
\description{
Normalize bin counts.
}
\author{
Jean Monlong
}

