% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair.discordant.reads.R
\name{pair.discordant.reads}
\alias{pair.discordant.reads}
\title{Pair discordant reads}
\usage{
pair.discordant.reads(chr, start.pos, end.pos, files.df, samp, controls,
  bins.df = NULL, plot = TRUE, nb.cores = 1)
}
\arguments{
\item{chr}{the chromosome name}

\item{start.pos}{the start position}

\item{end.pos}{the end position}

\item{files.df}{a data.frame with the path to the bam files for each samples.}

\item{samp}{the name of sample of interest.}

\item{controls}{the names of the samples to use as control.}

\item{bins.df}{a data.frame with the bins to use. If NULL (default), 1 Kbp bins will be used.}

\item{plot}{should a graph be displayed. Default is TRUE.}

\item{nb.cores}{the number of cores to use. Default is 1.}
}
\value{
a data.frame with the mapping summary for each pair of region.
}
\description{
Compare the pair mapping between regions of the genome. The goal is to assess/detect if two distant regions are linked by more read pairs than expected (in the controls).
}
\author{
Jean Monlong
}

