% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pca.norm}
\alias{pca.norm}
\title{PCA-based normalization of bin counts}
\usage{
pca.norm(bc, nb.pcs = 3, ref.samples = NULL)
}
\arguments{
\item{bc}{bin count matrix or data.frame (bins x samples)}

\item{nb.pcs}{the number of Principal Components to include in the regression model.}

\item{ref.samples}{the names of the reference samples on which the PCA is performed.
By default all samples are used.}
}
\value{
a matrix with the normalized bin counts (bin x sample)
}
\description{
Bin counts are normalized by regressing out the effect of the first Principal
Components. Beforehands, the average coverage is normalized. Then PC are computed
using \code{prcomp} function and regressed out using linear regression.
}
\author{
Jean Monlong
}

