\name{PopSV-package}
\alias{PopSV-package}
\alias{PopSV}
\docType{package}
\title{Population-based detection of structural variants from Read-Depth signal.}
\description{Detect abnormal read coverage using a set of samples as references. From high-throughput sequencing, the genome is binned and read mapping to each bin are counted. From these read counts, PopSV test, in each bin, is the read counts are different between a sample of interest and a set of reference samples. Functions for pre/post-processing are also provided.
}
\details{
\tabular{ll}{
Package: \tab PopSV\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-09-27\cr
License: \tab What license is it under? Good question\cr
}
 An overview of how to use the package, including the most important 
 functions 
}
\author{Jean Monlong <jean.monlong@mail.mcgill.ca>}
\references{
 Literature or other references for background information 
}
\keyword{structural variants, sequencing}
\seealso{
 Optional links to other man pages, e.g. 
}
\examples{}
