% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qc.sample.R
\name{qc.sample}
\alias{qc.sample}
\title{QC samples}
\usage{
qc.sample(bins.df, files.df = NULL, cnv.df = NULL, ref.samples = NULL,
  n.subset = 1000, nb.cores = 1)
}
\arguments{
\item{bins.df}{a data.frame with information about the bins. Columns 'chr', 'start', 'end' are required.}

\item{files.df}{a data.frame with information about the files, usually created by 'init.filenames'.}

\item{cnv.df}{a data.frame with PopSV calls, created by 'call.abnormal.cov'.}

\item{ref.samples}{a vector with the names of the samples used as reference. If NULL (default), all samples in 'files.df' are used.}

\item{n.subset}{the number of bins to use to compute the metrics. Default is 1000.}

\item{nb.cores}{number of cores to use. If higher than 1, \code{parallel}
package is used to parallelized the counting.}
}
\value{
a data.frame with qc metrics for each sample of the input 'files.df'.
}
\description{
Compute QC metrics for each sample informing how much it fits the reference samples. It can compute: the average distance between a sample and the 10% closest reference samples; the proportion of single-bin calls; the proportion of calls with copy number estimate close to 2. These metrics might be useful to interpret the calls. For example, an excess of calls in a sample which is globally an outlier in those QC is not so surprising.
}
\author{
Jean Monlong
}

