% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qc.samples.cluster.R
\name{qc.samples.cluster}
\alias{qc.samples.cluster}
\title{QC sample : clustering}
\usage{
qc.samples.cluster(bc.df, nb.rand.bins = NULL)
}
\arguments{
\item{bc.df}{a data.frame with the bin counts.}

\item{nb.rand.bins}{if non-NULL, the number of bins to randomly choose. Default is NULL.}
}
\value{
a shiny app in the web browser
}
\description{
Interactive representation of the bin counts with potential clustering of the samples.
}
\author{
Jean Monlong
}

