% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.zshift.unaffected.R
\name{qc.zshift.unaffected}
\alias{qc.zshift.unaffected}
\title{QC: Z-score bias in unaffected samples}
\usage{
qc.zshift.unaffected(files.df, bins.df, nb.cores = 1)
}
\arguments{
\item{files.df}{a data.frame with the path to relevant files for each samples.}

\item{bins.df}{a data.frame with information on which bins to analyze.}

\item{nb.cores}{the number of cores to use. Default is 1.}
}
\value{
a data.frame with a shift measure for each bin.
}
\description{
Checks that the Z-scores of unaffected samples are centered in 0. When controls are not perfectly matched, the Z-scores of cases can be slightly shifted and produce false calls. This function computes a measure of such bias.
}
\author{
Jean Monlong
}
