% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/quantile.norm.R
\name{quantile.norm}
\alias{quantile.norm}
\title{Quantile normalization}
\usage{
quantile.norm(bc.df, nb.cores = 1, norm.stats.comp = TRUE)
}
\arguments{
\item{bc.df}{a data.frame with 'chr', 'start', 'end' columns and then one column per sample with its bin counts.}

\item{nb.cores}{the number of cores to use. Default is 1.}

\item{norm.stats.comp}{Should some statistics on the normalized bin count be computed (mean, sd, outliers). Default is TRUE.}
}
\value{
a list with
\item{norm.stats}{a data.frame witht some metrics about the normalization of each
bin (row) : coverage average and standard deviation; number of outlier samples}
\item{bc.norm}{a data.frame, similar to the input 'bc.df', with the normalized bin counts.}
}
\description{
Quantile normalization of the bin counts. After such normalization, each sample has exactly the same global bin count distribution.
}
\author{
Jean Monlong
}

