% Generated by roxygen2 (4.0.2): do not edit by hand
\name{quick.count}
\alias{quick.count}
\title{Counts reads across samples in a small number of bins}
\usage{
quick.count(files.df, bins.df)
}
\arguments{
\item{files.df}{a data.frame with the information about the files to
use. Columns 'sample' and 'bam' are required.}

\item{bins.df}{a data.frame with the information about the bins. Columns 'chr', 'start'
and 'end' are required.}
}
\value{
a data.frame with the bin location and read counts for all samples.
}
\description{
This function retrieves bin count for all samples in the input file information
data.frame. It was designed to count reads in a small amount of bins (a subset
of what will be used for the full analysis). This counts can be used later to
compare a potentially large number samples and guide the selection of samples to consider in the analysis.
}
\author{
Jean Monlong
}

