% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.bedix.R
\name{read.bedix}
\alias{read.bedix}
\title{Retrieve a subset of an indexed BED file.}
\usage{
read.bedix(file, subset.reg = NULL, header = TRUE, as.is = TRUE,
  exact.match = FALSE)
}
\arguments{
\item{file}{the name of the file which the data are to be read from.
Should be compressed with |code{bgzip} and indexed by tabix algorithm.}

\item{subset.reg}{a data.frame or GRanges object with the regions to subset from.
If a data.frame, it must have columns named 'chr', 'start' and 'end'.}

\item{header}{Should the first line be used as headers. Default is TRUE.}

\item{as.is}{controls the conversion of columns. Default is TRUE, i.e. columns are
converted into 'character', 'numeric', 'integer', etc. If FALSE, some columns might
be converted into factors (e.g. 'character' columns).}

\item{exact.match}{Should the input bins be exactly matched in the output file. If FALSE (default), overlapping bins in 'file' will also be returned.}
}
\value{
a data.frame with the retrieved BED information.
}
\description{
Using \code{Rsamtools} functions to manipulate tabix file, the indexed BED file is read
and only regions defined by the desired subset are retrieved.
}
\author{
Jean Monlong
}

