% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/samples.merge.R
\name{samples.merge}
\alias{samples.merge}
\title{Merge data.frame across samples}
\usage{
samples.merge(files.df, samples, files.col, nb.cores = 1, sig.digits = 6)
}
\arguments{
\item{files.df}{a data.frame with information about each output file.}

\item{samples}{the names of the samples to use.}

\item{files.col}{the name of 'files.df' to use for each 'res' element.}

\item{nb.cores}{the number of cores to use.}

\item{sig.digits}{the number of significant digits. Default is 6.}
}
\value{
the merged data.frame
}
\description{
Merge files for different samples.
}
\author{
Jean Monlong
}

