% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.sv.interactive}
\alias{summary.sv.interactive}
\title{Interactive summary}
\usage{
\method{summary}{sv.interactive}(res.df)
}
\arguments{
\item{res.df}{the data.frame with the results.}
}
\value{
a shiny app in the web browser.
}
\description{
Interactive summary of the individual bins with abnormal coverage as detected per 'call.abnomal.cov'.
}
\author{
Jean Monlong
}

