% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sv.summary.R
\name{sv.summary}
\alias{sv.summary}
\title{Abnormal regions summary}
\usage{
sv.summary(res.df, out.pdf = NULL, print = TRUE)
}
\arguments{
\item{res.df}{the data.frame with the results.}

\item{out.pdf}{the name for the output PDF file to create. If NULL (default), the graphs are either displayed directly or returned as a list.}

\item{print}{Should the graphs be displayed ? If 'out.pdf' though, this parameter is not used as no graph will be displayed.}
}
\value{
A list with :
\item{graphs.l}{a list of ggplot graphs}
\item{info.df}{a data.frame with different summary statistics}
}
\description{
Summary of regions with abnormal coverage as detected by 'call.abnomal.cov'.
}
\author{
Jean Monlong
}

