% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmm.norm.R
\name{tmm.norm}
\alias{tmm.norm}
\title{Trimmed-Mean M normalization}
\usage{
tmm.norm(
  bc.df,
  cont.sample,
  trim.level = 0.3,
  nb.cores = 1,
  norm.stats.comp = TRUE
)
}
\arguments{
\item{bc.df}{a data.frame with 'chr', 'start', 'end' columns and then one column per sample with its bin counts.}

\item{cont.sample}{the sample to use as baseline for the pairwise normalization.
All the samples will be normalized to it.}

\item{trim.level}{How much of the M values should be trimmed (on each side of the distribution). Default is 0.3 (30\%).}

\item{nb.cores}{the number of cores to use. Default is 1.}

\item{norm.stats.comp}{Should some statistics on the normalized bin count be computed (mean, sd, outliers). Default is TRUE.}
}
\value{
a list with
\item{norm.stats}{a data.frame witht some metrics about the normalization of each
bin (row) : coverage average and standard deviation; number of outlier samples}
\item{bc.norm}{a data.frame, similar to the input 'bc.df', with the normalized bin counts.}
}
\description{
TMM normalization of bin counts. After such normalization, the majority of bins have the same trimmed average across samples.
}
\author{
Jean Monlong
}
