% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tn.norm.qc.div.R
\name{tn.norm.qc.div}
\alias{tn.norm.qc.div}
\title{Targeted normalization QC: diversity of the supporting bins}
\usage{
tn.norm.qc.div(norm.stats, out.pdf = "normStats-QC-supportDiversity.pdf",
  chunk.size = 1000, nb.cores = 1)
}
\arguments{
\item{norm.stats}{the name of the file with the normalization statistics ('norm.stats' in 'tn.norm' function) or directly a 'norm.stats' data.frame.}

\item{out.pdf}{the name of the PDF file to create.}

\item{chunk.size}{the size of a chunk to read the file. Default is 1000.}

\item{nb.cores}{the number of cores to use. Default is 1.}
}
\value{
a data.frame with the bins location and the computed metric.
}
\description{
Computes the diversity of the supporting bins used during normalization. For now it only computes the number of bins in the same chromosome as the bin to normalize. To avoid over-normalization of large CNVs or full-chromosome aneuploidy (e.g. in cancer), we would like the majority of the supporting to be located different chromosomes than the bin to normalize.
}
\author{
Jean Monlong
}
