% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tn.norm.qc}
\alias{tn.norm.qc}
\title{QC graphs for Targeted Normalization}
\usage{
tn.norm.qc(norm.stats, out.pdf = "normStats-QC.pdf")
}
\arguments{
\item{norm.stats}{the name of the file with the normalization statistics ('norm.stats' in 'tn.norm' function) or directly a 'norm.stats' data.frame.}

\item{out.pdf}{the name of the PDF file to create.}
}
\value{
the name of the created PDF file.
}
\description{
Following targeted normalization, some Quality Control graphs are produced to see if some bins were problematic. TODO MORE
}
\author{
Jean Monlong
}

