% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tn.norm.R
\name{tn.norm}
\alias{tn.norm}
\title{Targeted-normalization of bin counts}
\usage{
tn.norm(bc, cont.sample, nb.support.bins = 1000, bins = NULL,
  save.support.bins = TRUE, bootstrap = FALSE)
}
\arguments{
\item{bc}{a matrix or data.frame with the bin counts (bin x sample).}

\item{cont.sample}{the sample to use as baseline for the pairwise normalization.
All the samples will be normalized to it.}

\item{nb.support.bins}{the number of bins to use for the normalization.}

\item{bins}{a vector the names of the bins to normalize. If NULL (default), all
bins are normalized.}

\item{save.support.bins}{if TRUE (default) the bins used for the normalization are
saved in the output object 'norm.stats'.}

\item{bootstrap}{should the supporting bins defined using a bootstrap approach. Default is FALSE.}
}
\value{
a list with
\item{norm.stats}{a data.frame witht some metrics about the normalization of each
bin (row) : correlation with worst supporting bin; coverage average and standard
deviation; number of outlier reference samples; supporting bins.}
\item{bc.norm}{a matrix with the normalized bin counts (bin x sample).}
\item{nb.support.bins, cont.sample, z.poisson}{a backup of the input parameters.}
\item{cont.sample}{the name of the sample used to normalize all samples.}
}
\description{
Bin counts are normalized one bin at a time, using a subset of the bins that look
similar across the reference samples.
}
\author{
Jean Monlong
}

