% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tn.test.sample.R
\name{tn.test.sample}
\alias{tn.test.sample}
\title{Single sample targeted normalization and test}
\usage{
tn.test.sample(test.sample, files.df, cont.sample, bc.ref.f, norm.stats.f,
  write.out.file = TRUE, compress.index = TRUE, z.poisson = FALSE,
  aberrant.cases = FALSE, append = FALSE)
}
\arguments{
\item{test.sample}{the name of the sample to test.}

\item{files.df}{a data.frame with the information about the files to
use. Columns 'sample' and 'bc.gc.bg' are required and should be
present  after running 'initFileNames' function. Files should exist if
'correct.GC' was run.}

\item{cont.sample}{the name of the sample used as control for the normalization.}

\item{bc.ref.f}{the path to the input file used for targeted normalization ('tn.norm').}

\item{norm.stats.f}{the name of the file with the statistic of the targeted normalization run.}

\item{write.out.file}{should the result be written in files (from 'z' and 'fc' columns in 'files.df'). Default is TRUE.}

\item{compress.index}{should the output files be compressed and indexed. Default is TRUE.}

\item{z.poisson}{Should the Z-score be computed as an normal-poisson hybrid (see
Details). Default is FALSE.}

\item{aberrant.cases}{if TRUE (default) a more robust (but sligthly longer) normalization
is performed on cases to deal with potential large chromosomal aberrations. In practice,
it is recommended for cancer but can be turned off if less than ~20% of the genome is expected
to be affected.}

\item{append}{should the results be appended to existing files. Default is FALSE.}
}
\value{
a data.frame with columns :
\item{chr, start, end}{the location of the bin}
\item{bc}{the normalized bin count}
\item{z}{the Z-scores}
\item{fc}{the fold-change compared to the average bin count in the reference samples}
}
\description{
Bin counts from one sample are normalized following instructions from
a previous targeted-normalization run.
}
\details{
The Z-score is computed by substracting the bin count by the average bin count
across the reference samples and dividing by their standard deviation. If
'z.poisson' is TRUE, a score using Poisson distribution is also computed, using
the average bin count as an estimator of the lambda. Then the score with the lowest
absolute value is kept. This hybrid Z-score is to be used when some regions have low
coverage where it is more robust to use Poisson assumptions.
}
\author{
Jean Monlong
}

