% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/winsor.R
\name{winsor}
\alias{winsor}
\title{Winsorize a vector}
\usage{
winsor(x, u = 5, l = NULL)
}
\arguments{
\item{x}{the input to winsorize.}

\item{u}{the highest value desired.}

\item{l}{the lowest value desired.}
}
\value{
a vector with winsorized values.
}
\description{
Winsorising a vector: any value higher(lower) than U(L) are changed to U(L).
}
\author{
Jean Monlong
}
\keyword{internal}

