% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{wrap}
\alias{wrap}
\title{Wrapper used for making pipelines}
\usage{
wrap(args = commandArgs(TRUE))
}
\arguments{
\item{args}{command line arguments}
}
\value{
the status of the computation.
}
\description{
Meant to be used from the command line with 'Rscript "PopSV::wrap()" countsample samp1 config.RData bins.RData'.
}
\author{
Jean Monlong
}
