% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write.split.samples.R
\name{write.split.samples}
\alias{write.split.samples}
\title{Split and write results in one file per sample}
\usage{
write.split.samples(res, files.df, samples, res.n = c("z", "fc"),
  files.col = c("z", "fc"), compress.index = TRUE, nb.cores = 1,
  append = FALSE)
}
\arguments{
\item{res}{a list with the results to split/write.}

\item{files.df}{a data.frame with information about each output file.}

\item{samples}{the names of the samples to use.}

\item{res.n}{the name of 'res' elements containing joined results.}

\item{files.col}{the name of 'files.df' to use for each 'res' element.}

\item{compress.index}{should the output be compressed and indexed. Default is TRUE.}

\item{nb.cores}{the number of computing cores to use. Default is 1.}

\item{append}{should the results be appended at the end of existing files. Default is FALSE.}
}
\value{
'Done' if everything worked fine.
}
\description{
Split join results into one file per sample per feature. For example to split and write the Z-scores of reference samples after joined computation.
}
\author{
Jean Monlong
}

