% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeVcf.R
\name{writeVcf}
\alias{writeVcf}
\title{Write VCF file}
\usage{
writeVcf(cnv.df, output.file = "cnvs.vcf",
  genome = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19)
}
\arguments{
\item{cnv.df}{the data.frame with the cnv calls (from 'call.abnormal.cov').}

\item{output.file}{the file name for the output VCF file. Default is 'cnvs.vcf'}

\item{genome}{BSgenome to retrieve REF nucleotide. Default is hg19 (BSgenome.Hsapiens.UCSC.hg19::Hsapiens).}
}
\value{
the output file name.
}
\description{
Transforms the BED/data.frame format into a VCF format and write a VCF file.
The calls are merged if they have exactly the same coordinates ('chr', 'start' and 'end').
}
\author{
Jean Monlong
}
