% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/z.comp.R
\name{z.comp}
\alias{z.comp}
\title{Z-score computation}
\usage{
z.comp(files.df, samples, msd.f = NULL, z.poisson = FALSE,
  col = "bc.gc.norm.gz", nb.cores = 1)
}
\arguments{
\item{files.df}{a data.frame with the file paths.}

\item{samples}{the samples to analyze.}

\item{msd.f}{the path to the file with mean/sd bin count on reference samples.}

\item{z.poisson}{Should the Z-score be computed as an normal-poisson hybrid (see
Details). Default is FALSE.}

\item{col}{the column in 'files.df' that define the bin count file path.}

\item{nb.cores}{the number of cores to use.}
}
\value{
a list with
\item{z}{a data.frame with the Z-scores for each bin and sample (bin x sample).}
\item{fc}{a data.frame with the fold-change compared to the average bin count in
the reference samples for each bin and sample (bin x sample).}
\item{msd}{the mean, standard deviation and number of removed outlier samples in each bin.}
}
\description{
Z-score computation from bin count and/or mean/sd metrics on the reference samples
}
\author{
Jean Monlong
}

