% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/z.norm.R
\name{z.norm}
\alias{z.norm}
\title{Z-score normalization}
\usage{
z.norm(z.df, gc.df, class.size = 5000)
}
\arguments{
\item{z.df}{a data.frame with the Z-scores.}

\item{gc.df}{a data.frame with the GC content.}

\item{class.size}{the number of bins in each GC classes. Default is 5000.}
}
\value{
a vector with the normalized Z-scores.
}
\description{
The Z-score is normalized to ensure centered distribution and similar variance across different GC classes. This is usually not necessary when the quality is comparable across the samples. However if some samples contain experiment-specific biases this approach can reduce false calls.
}
\author{
Jean Monlong
}
\keyword{internal}

