% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aneuploidy.flag.R
\name{aneuploidy.flag}
\alias{aneuploidy.flag}
\title{Flag chromosomal aneuploidy}
\usage{
aneuploidy.flag(files.df, col.file = "bc.gc.gz", verbose = TRUE,
  ref.samples = NULL, centromere.pos = NULL, nb.cores = 1)
}
\arguments{
\item{files.df}{a data.frame with information about samples such as the path to the count files.}

\item{col.file}{the name of the column in 'files.df' with the path information to use. Default is 'bc.gc.gz', i.e. GC corrected bin counts.}

\item{verbose}{Shoulf progress be outputted ? Default is TRUE.}

\item{ref.samples}{a vector with the names of the samples to be used as reference (e.g. normals). If NULL (default) all samples are used.}

\item{centromere.pos}{a vector with the position of the centromeres for each chromosome. Each position should be named with the chromosome name. If NULL (default), full chromosome aneuploidy is tested instead of arm-level deletion/duplication.}

\item{nb.cores}{the number of processing cores to use, Default is 1.}
}
\value{
a data.frame with columns:
\item{sample}{the sample}
\item{chr}{the chromosome}
\item{pv.loss, pv.grain}{the p-value for a loss/gain}
\item{loss/gain}{is the loss/gain significant}
\item{bc.norm}{the normalized coverage}
\item{exp.bc}{the expected normalized coverage}
\item{bc.diff}{the difference between observed and expected coverage}
}
\description{
Chromosomes with clear aneuploidy are detected using a simple threshold-based approach. It is used to flag chromosome with complete aneuploidy that could handicap later analysis. However it might not detect partial chromosomal aberration or aneuploidy in samples with noisy read coverage.
}
\author{
Jean Monlong
}

