% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aneuploidy.flag.R
\name{aneuploidy.flag}
\alias{aneuploidy.flag}
\title{Flag chromosomal aneuploidy}
\usage{
aneuploidy.flag(samp, files.df, col.file = "bc.gz", nb.bins = 1000,
  prop.aneu = 0.1, plot = FALSE)
}
\arguments{
\item{samp}{the name of the sample to analyze.}

\item{files.df}{a data.frame with information about samples such as the path to the count files.}

\item{col.file}{the name of the column in 'files.df' with the path information to use.}

\item{nb.bins}{the number of bins to use when subsampling each chromosome. Default is 1000.}

\item{prop.aneu}{the proportion of normal bins under which a chromosome is considered aneuploid.}

\item{plot}{should the read coverage per chromosome be displayed. Default is FALSE.}
}
\value{
a list with:
\item{aneu.chrs}{vector with the names of the flagged(aneuploid) chromosomes.}
\item{aneu.chrs.fc}{the fold change in read count in the aneuploid chromosomes.}
}
\description{
Chromosomes with clear aneuploidy are detected using a simple threshold-based approach. It is used to flag chromosome with complete aneuploidy that could handicap later analysis. It might not detect partial chromosomal aberration or aneuploidy in samples with noisy read coverage.
}
\author{
Jean Monlong
}

