% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/breakpoint.finder.interactive.R
\name{breakpoint.finder.interactive}
\alias{breakpoint.finder.interactive}
\title{Interactive breakpoint finder}
\usage{
breakpoint.finder.interactive(chr, start, end, test.sample, files.df,
  ref.samples, proper = TRUE, nb.cores = 1, bp.res = 1)
}
\arguments{
\item{chr}{the chromosome}

\item{start}{the start position}

\item{end}{the end position}

\item{test.sample}{the sample with abnormal coverage}

\item{files.df}{a data.frame with the path to the bam files}

\item{ref.samples}{other samples to visualize}

\item{proper}{should proper mapping be counted.}

\item{nb.cores}{number of cores}

\item{bp.res}{default resolution (bp).}
}
\value{
a data.frame with the breakpoint coordinates
}
\description{
Interactive tool to fine tune breakpoint location
}
\author{
Jean Monlong
}

