% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/breakpoint.finder.interactive.R
\name{breakpoint.finder.interactive}
\alias{breakpoint.finder.interactive}
\title{Interactive breakpoint finder}
\usage{
breakpoint.finder.interactive(chr, start, end, test.sample, files.df,
  ref.samples, proper = TRUE, nb.cores = 1, bp.res = 1, flanks = 2000,
  related.samples = NULL)
}
\arguments{
\item{chr}{the chromosome}

\item{start}{the start position}

\item{end}{the end position}

\item{test.sample}{the sample with abnormal coverage}

\item{files.df}{a data.frame with the path to the bam files}

\item{ref.samples}{other samples to visualize}

\item{proper}{should proper mapping be counted.}

\item{nb.cores}{number of cores}

\item{bp.res}{default resolution (bp)}

\item{flanks}{default flank size (bp)}

\item{related.samples}{samples related to the `test.sample` (e.g. parents) to be drawn in a special color.}
}
\value{
a list with
\item{bk.df}{a data.frame with the breakpoint coordinates}
\item{graph}{the ggplot object with the final graph}
}
\description{
Interactive tool to fine tune breakpoint location. An interactive Shiny app will open in your web browser. There the bp coverage of the `test.sample` will be shown along with the coverage of the reference samples (`ref.samples`). The user can then slide the breakpoint where the tested samples starts to deviate from the reference. If the deviation is not clear a few parameters (e.g. mapping quality, resolution) can be changed live. Once satisfied the user pushes a button and the function returns the new breakpoint locations. In practice this function can be looped over to quickly fine-tune the breakpoints of several CNVs.
}
\author{
Jean Monlong
}

