% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint.finder.R
\name{breakpoint.finder}
\alias{breakpoint.finder}
\title{Breakpoint finder (NOT READY YET)}
\usage{
breakpoint.finder(range.df, files.df, test.sample, ref.samples, proper = TRUE,
  map.quality = 0, bp.res = 3:6, slidW.step = 1, midread = FALSE,
  nb.cores = 1)
}
\arguments{
\item{range.df}{a data.frame with the information about the genomic ranges to analyze.}

\item{files.df}{a data.frame with information about the BAM file path for each sample (columns 'sample' and 'bam').}

\item{test.sample}{the name of the sample of interest.}

\item{ref.samples}{the names of the samples to use as reference for the breakpoint detection.}

\item{proper}{Should proper mapping be counted. Default is TRUE.}

\item{map.quality}{the minimum mapping quality for the counted reads. Default is 30.}

\item{bp.res}{the base-pair resolution of the estimation.}

\item{slidW.step}{the step used when sliding the window. Default is 1.}

\item{midread}{should the middle of the read used to asses coverage. If FALSE (default), the entire read counts for coverage computation.}

\item{nb.cores}{the number of processing cores to use, Default is 1.}
}
\value{
the original data.frame with updated breakpoint location.
}
\description{
An attempt to find breakpoints as the location with the most dramatic local coverage differences between a samples and a set of reference samples. DON'T TRUST THIS FUNCTION YET.
}
\author{
Jean Monlong
}

