% Generated by roxygen2 (4.0.2): do not edit by hand
\name{call.abnormal.cov}
\alias{call.abnormal.cov}
\title{Call abnormal bins}
\usage{
call.abnormal.cov(z, samp, out.pdf = NULL, FDR.th = 0.05,
  merge.cons.bins = c("stitch", "zscores", "no"), z.th = c("sdest",
  "consbins"), fc = NULL, norm.stats = NULL, d.max.max = 0.5,
  min.normal.prop = 0.5, aneu.chrs = NULL, ref.dist.weight = NULL)
}
\arguments{
\item{z}{the name of the file with the Z-scores for all samples OR or a
data.frame with the Z-scores for all samples.}

\item{samp}{the name of the sample to analyze.}

\item{out.pdf}{the name of the output pdf file.}

\item{FDR.th}{the False Discovery Rate to use for the calls. Further filtering
can always be performed.}

\item{merge.cons.bins}{how the bins should be merged. Default is 'stitch'. 'zscores' is another approch (see Details), 'no' means no bin merging.}

\item{z.th}{how the threshold for abnormal Z-score is chosen. Default is 'sdest' which will use 'FDR.th=' parameter as well. 'consbins' looks at the number of consecutive bins, see Details.}

\item{fc}{the name of the file with the copy number estimates for all samples OR
or a data.frame with the copy number estimates for all samples.}

\item{norm.stats}{the name of the file with the normalization statistics ('norm.stats' in 'tn.norm' function) or directly a 'norm.stats' data.frame.}

\item{d.max.max}{the maximum correlation of the last supporting bin.}

\item{min.normal.prop}{the minimum proportion of the regions expected to be normal. Default is 0.5. For cancers with many large aberrations, this number can be lowered.}

\item{aneu.chrs}{the names of the chromosomes to remove because flagged as aneuploid. If NULL (default) all chromosomes are analyzed.}

\item{ref.dist.weight}{the weight (value between 0 and 1) based on the distance to the reference samples.}
}
\value{
a data.frame with columns
\item{chr, start, end}{the genomic region definition}
\item{z}{the Z-score}
\item{pv, qv}{the P-value and Q-value(~FDR)}
\item{fc}{the copy number estimate (if 'fc' was not NULL).}
\item{nb.bin.cons}{the number of consecutive bins (if the bins were merged;
'merge.cons.bins!="no"')}
\item{cn2.dev}{Copy number deviation from the reference }
}
\description{
Detect abnormal bin from the Z-score distribution. A normal distribution is first
fitted to the Z-score distribution. P-values are computed from this estimated
null distribution and corrected for multiple testing. Eventually consecutive bins
with abnormal read counts can be merged.
}
\details{
DETAILS ON THE MERGING APPROACH

DETAILS ON THE NORMAL FITTING AND Z THRESHOLDING
}
\author{
Jean Monlong
}

