% Generated by roxygen2 (4.0.2): do not edit by hand
\name{call.abnormal.cov}
\alias{call.abnormal.cov}
\title{Call abnormal bins}
\usage{
call.abnormal.cov(z, samp, out.pdf = NULL, FDR.th = 0.05,
  merge.cons.bins = FALSE, fc = NULL, norm.stats = NULL,
  d.max.max = 0.5, min.normal.prop = 0.5)
}
\arguments{
\item{z}{the name of the file with the Z-scores for all samples OR or a
data.frame with the Z-scores for all samples.}

\item{samp}{the name of the sample to analyze.}

\item{out.pdf}{the name of the output pdf file.}

\item{FDR.th}{the False Discovery Rate to use for the calls. Further filtering
can always be performed.}

\item{merge.cons.bins}{Should consecutive abnormal bins be merged. If FALSE (default)
single bin calls are reported. If TRUE, reported bins are merged.}

\item{fc}{the name of the file with the copy number estimates for all samples OR
or a data.frame with the copy number estimates for all samples.}

\item{norm.stats}{the name of the file with the normalization statistics ('norm.stats' in 'tn.norm' function) or directly a 'norm.stats' data.frame.}

\item{d.max.max}{the maximum correlation of the last supporting bin.}

\item{min.normal.prop}{the minimum proportion of the regions expected to be normal. Default is 0.5. For cancers with many large aberrations, this number can be lowered.}
}
\value{
a data.frame with columns
\item{chr, start, end}{the genomic region definition}
\item{z}{the Z-score}
\item{pv, qv}{the P-value and Q-value(~FDR)}
\item{fc}{the copy number estimate (if 'fc' was not NULL).}
\item{nb.bin.cons}{the number of consecutive bins (if the bins were merged;
'merge.cons.bins=TRUE')}
\item{cn2.dev}{Copy number deviation from the reference }
}
\description{
Detect abnormal bin from the Z-score distribution. A normal distribution is first
fitted to the Z-score distribution. P-values are computed from this estimated
null distribution and corrected for multiple testing. Eventually consecutive bins
with abnormal read counts can be merged.
}
\details{
DETAILS ON THE MERGING APPROACH

DETAILS ON THE NORMAL FITTING
}
\author{
Jean Monlong
}

