% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comp.index.files.R
\name{comp.index.files}
\alias{comp.index.files}
\title{Compress and index BED-like files}
\usage{
comp.index.files(files, outprefix = files, rm.input = TRUE,
  overwrite.out = TRUE, reorder = FALSE)
}
\arguments{
\item{files}{the names of the files to compress and index.}

\item{outprefix}{the prefix to use when naming the output files. By default, the original file names.}

\item{rm.input}{should the original input file be removed after compression/indexing ? Default is TRUE.}

\item{overwrite.out}{should a compressed file be overwritten if it already exists ? Default is TRUE.}

\item{reorder}{should the files be read and reordered before compression/indexation. Default is FALSE.}
}
\value{
a character vector confirming the creation of the new files.
}
\description{
Compress and index BED-like files. The first three columns must define chromosome location, start and end position.
}
\details{
The files should be ordered by genomic position. If not 'reorder=TRUE' will force the files to be imported in R, reordered and written back.
}
\author{
Jean Monlong
}

