% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEmptyDF.R
\name{createEmptyDF}
\alias{createEmptyDF}
\title{Create empty data.frame}
\usage{
createEmptyDF(colClasses, nrows)
}
\arguments{
\item{colClasses}{a vector with the types definition for the columns.}

\item{nrows}{the number of rows in the data.frame to create.}
}
\value{
a data.frame with 'NA' but of the specified types
}
\description{
Create an empty data.frame with NA and specified column types. Probably not the best way to do.
}
\author{
Jean Monlong
}
\keyword{internal}

