% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fdrtool.quantile}
\alias{fdrtool.quantile}
\title{P-values estimation}
\usage{
fdrtool.quantile(z, quant.int = seq(0.2, 1, 0.02))
}
\arguments{
\item{z}{a vector with the Z-scores}

\item{quant.int}{a vector with the quantiles to test for the estimation of the
null normal variance.}
}
\value{
a list with
\item{pval}{the vector of P-values}
\item{qval}{the vector of Q-values / FDR estimates}
\item{quant.int}{the quantile used for the estimation of the null distribution
variance.}
\item{sigma.est}{the estimated null distribution variance}
}
\description{
Compute P-values and Q-values from the Z-score distribution.
}
\author{
Jean Monlong
}
\keyword{internal}

