% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fragment.genome.hp19}
\alias{fragment.genome.hp19}
\title{Fragment hg19 genome}
\usage{
fragment.genome.hp19(bin.size = 1000)
}
\arguments{
\item{bin.size}{the size of the bins}
}
\value{
a data.frame with columns 'chr', 'start' and 'end'.
}
\description{
Fragment hg19 genome into consecutive bins of equal size. This function
require package \code{BSgenome.Hsapiens.UCSC.hg19} to be installed.
}
\author{
Jean Monlong
}

