% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/freq.range.R
\name{freq.range}
\alias{freq.range}
\title{Frequency computation for ranges}
\usage{
freq.range(range.df, plot = FALSE)
}
\arguments{
\item{range.df}{a data.frame with columns 'chr', 'start' and 'end'.}

\item{plot}{should a graph with the frequency distribution be displayed.}
}
\value{
a data.frame with the frequency of each sub-range
}
\description{
All the ranges defined by columns 'chr', 'start' and 'end' in the input data.frame are overlapped and used to define sub-ranges. Then the number of input ranges overlapping each sub-ranges is computed and return in a data.frame.
}
\author{
Jean Monlong
}

