% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.range.R
\name{freq.range}
\alias{freq.range}
\title{Frequency computation for ranges}
\usage{
freq.range(range.df, plot = FALSE, annotate.only = FALSE,
  chunk.size = 10000)
}
\arguments{
\item{range.df}{a data.frame with columns 'chr', 'start' and 'end'.}

\item{plot}{should a graph with the frequency distribution be displayed. Default is FALSE.}

\item{annotate.only}{If TRUE, the input ranges are annotated with the number of other overlapping ranges. If FALSE (Default), the input ranges are fragmented and the frequency computed for each sub-range.}

\item{chunk.size}{the size of the chunk to analyze. Useful when many ranges and/or when 'annotate.only=TRUE'. Default is 10000.}
}
\value{
a data.frame with the frequency of each sub-range
}
\description{
The frequency is defined as the number of input ranges overlapping a particular bp/region. In practice, all the ranges defined by columns 'chr', 'start' and 'end' in the input data.frame are overlapped and used to define sub-ranges. Then the number of input ranges overlapping each sub-ranges is computed and return in a data.frame.
}
\details{
If 'annotate.only=TRUE' however, the frequency of an input range is computed as the number of ranges overlapping it. It is less-suited to describe the frequency distribution, but more convenient to filter out frequent variants.
}
\author{
Jean Monlong
}
