% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGC.R
\name{getGC}
\alias{getGC}
\title{GC content computation for specific bins}
\usage{
getGC(bins.df, genome = "hg19")
}
\arguments{
\item{bins.df}{a data.frame with bin information. Columns 'chr', 'start' and 'end'
are required.}

\item{genome}{either 'hg19', 'GRCh38' or a BSgenome object with the genome.}
}
\value{
the same input data.frame with an extra column 'GCcontent'.
}
\description{
Add a column with GC content to a data.frame with 'chr', 'start' and 'end' column.
Requires BSgenome for Hsapiens hg19; if not instructions to install it will
be displayed.
}
\author{
Jean Monlong
}
