% Generated by roxygen2 (4.0.2): do not edit by hand
\name{medvar.norm}
\alias{medvar.norm}
\title{Median-variance normalization of bin counts}
\usage{
medvar.norm(bc)
}
\arguments{
\item{bc}{a matrix or data.frame with the bin counts (bin x sample).}
}
\value{
a matrix with the normalized bin counts (bin x sample).
}
\description{
The median and deviation to the median is normalized across samples.
}
\author{
Jean Monlong
}

