% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mergeConsBin.reduce.R
\name{mergeConsBin.reduce}
\alias{mergeConsBin.reduce}
\title{Merge nearby bins}
\usage{
mergeConsBin.reduce(df, col.mean = c("z", "pv", "qv", "fc"),
  stitch.dist = 10000)
}
\arguments{
\item{df}{a data.frame with at least 'chr', 'start' and 'end' columns.}

\item{col.mean}{the columns from which the average across the bins to merge will be computed.}

\item{stitch.dist}{the stitching distance, i.e. the maximum distance at which two bins will be merged.}
}
\value{
a data.frame, similar to the input but with merged rows.
}
\description{
Merge nearby bins using a user-defined 'stitching' distance. Duplication and deletions are stitched separately.
}
\author{
Jean Monlong
}

