% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.tm.opt.R
\name{norm.tm.opt}
\alias{norm.tm.opt}
\title{Trimmed-Mean normalization optimized}
\usage{
norm.tm.opt(df, ref.col, lm.min.prop = 0.1, bc.mean.norm = NULL,
  chrs = NULL)
}
\arguments{
\item{df}{a data.frame with the bin counts (bin x sample).}

\item{ref.col}{the column to be used as baseline for the pairwise normalization.
All columns are normalized on this sepcified column.}

\item{lm.min.prop}{the minimum height proportion in the density curve to define a
local maximum.}

\item{bc.mean.norm}{the average coverage in the reference sample. To be used to
help normalize tumor samples.}

\item{chrs}{the chromosomes of the bins used. The diversity of the chromosome is
useful to differentiate misleading coefficient from a few aberrant chromosomes.}
}
\value{
a vector with the normalization coefficients for each column (sample).
}
\description{
Optimized version of the trimmed-mean normalization for bin counts.
}
\author{
Jean Monlong
}
\keyword{internal}
