% Generated by roxygen2 (4.0.2): do not edit by hand
\name{normQC}
\alias{normQC}
\title{Normalized bin count QC metrics}
\usage{
normQC(bc.mat, n.subset = 10000)
}
\arguments{
\item{bc.mat}{matrix with bin counts (bins x samples).}

\item{n.subset}{number of bins to use for the analysis. Default is 10 000. Bins are selected randomly.}
}
\value{
a list with
\item{prop.non.normal.bin}{proportion of bins with non-normal distribution across samples.}
\item{prop.non.poisson.bin}{proportion of bins with non-Poisson distribution across samples.}
\item{prop.nonRand.rank}{proportion of bins with non-random ranks.}
\item{prop.non.norm.z.mean}{average (across samples) proportion of bins with non-random Z-scores.}
\item{prop.non.norm.z.max}{maximum (i.e for worst sample) proportion of bins with non-random Z-scores.}
\item{prop.non.norm.z}{proportion of bins with non-random Z-scores, for each sample.}
\item{n.subset}{number of bins used for the analysis.}
}
\description{
Quality control statistics of bin counts. First, the counts are tested to follow a
normal (or Poisson) distribution across samples, in each bin. Then, the randomness
of sample ranks are tested. Finally, Z-scores are computed for a subset of the bins
and their normality is tested. The second and third test are the most important.
Indeed, consistent rankings supports sample-specific technical bias, hence reduced
power to detect "true" abnormal read counts. Non-normal Z-scores will lead to
inappropriate fit for the null distribution.
}
\details{
Shapiro test is used to test normality of the bin counts across samples. The proportion
of bins with non-normal distribution is derived from the Pi0 estimate estimated by
package \code{qvalue}. Pi0 is the proportion of pvalues following the null distribution.

Goodness of fir from package \code{vcd} is used to test if bin counts follow a Poisson
distribution. Again Pi0 estimate from \code{qvalue} package is used to compute the proportion
of bin that don't follow Poisson distribution.

The randomness of the sample ranks are tested using a Chi-Square test. To have an
idea of the proportion of the bins with abnormal ranking, bins are randomly divided
into groups and the test performed in each group. The proportion of bin with non-random
ranks is approximated by the proportion of groups which failed the Chi-Square test.

Z-scores normality is computed by comparing their density distribution and a fitted
normal distribution. The estimate represents the proportion of the area under the curve
that is unique to the Z-score curve.
}
\author{
Jean Monlong
}

