% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.norm.R
\name{pca.norm}
\alias{pca.norm}
\title{PCA-based normalization of bin counts}
\usage{
pca.norm(bc.df, nb.pcs = 3, nb.cores = 1, norm.stats.comp = TRUE)
}
\arguments{
\item{bc.df}{a data.frame with 'chr', 'start', 'end' columns and then one column per sample with its bin counts.}

\item{nb.pcs}{the number of Principal Components to include in the regression model.}

\item{nb.cores}{the number of cores to use. Default is 1.}

\item{norm.stats.comp}{Should some statistics on the normalized bin count be computed (mean, sd, outliers). Default is TRUE.}
}
\value{
a list with
\item{norm.stats}{a data.frame witht some metrics about the normalization of each
bin (row) : coverage average and standard deviation; number of outlier reference samples; principal components}
\item{bc.norm}{a data.frame, similar to the input 'bc.df', with the normalized bin counts.}
}
\description{
Bin counts are normalized by regressing out the effect of the first Principal
Components. Beforehands, the average coverage is normalized. Then PC are computed
using \code{prcomp} function and regressed out using linear regression.
}
\author{
Jean Monlong
}
