% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopSV-package.R
\docType{package}
\name{PopSV-package}
\alias{PopSV-package}
\title{Population-based detection of structural variants from Read-Depth signal.}
\description{
Detect abnormal read coverage using a set of samples as references. From high-throughput sequencing, the genome is binned and read mapping to each bin are counted. From these read counts, PopSV test, in each bin, is the read counts are different between a sample of interest and a set of reference samples. Functions for pre/post-processing are also provided.
}
\details{
\tabular{ll}{
Package: \tab PopSV\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-12-09\cr
License: \tab GPL-2\cr
}
}
\seealso{
\url{http://jmonlong.github.io/PopSV/}
}
\author{
Jean Monlong \email{jean.monlong@mail.mcgill.ca}
}
