% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qc.samples.cluster.R
\name{qc.samples.cluster}
\alias{qc.samples.cluster}
\title{QC sample : clustering}
\usage{
qc.samples.cluster(bc.df, nb.rand.bins = NULL)
}
\arguments{
\item{bc.df}{a data.frame with the bin counts.}

\item{nb.rand.bins}{if non-NULL, the number of bins to randomly choose. Default is NULL.}
}
\value{
a shiny app in the web browser
}
\description{
Interactive representation of the bin counts with potential clustering of the samples. This app can be used to check if the reference samples are homogeneous, and define outliers or batches.
}
\details{
There is no need to use all the bin count, a few thousand bins is enough to detect potential problems. 'quick.count' is useful to retrieve bin counts on a subset of bins.
}
\author{
Jean Monlong
}

