% Generated by roxygen2 (4.0.2): do not edit by hand
\name{qc.samples}
\alias{qc.samples}
\title{Join and QC of samples}
\usage{
qc.samples(files.df, bin.df, ref.samples = NULL, outfile.prefix,
  out.pdf = NULL, appendIndex.outfile = TRUE, chunk.size = 1e+05,
  col.bc = "bc.gc.gz")
}
\arguments{
\item{files.df}{a data.frame with the information about the files to
use. Columns 'sample' and 'bc.gc.bg' are required and should be
present  after running 'initFileNames' function. Files should exist if
'correct.GC' was run.}

\item{bin.df}{a data.frame with the information about the bins. Columns 'chr', 'start'
and 'end' are required.}

\item{ref.samples}{a vector with the names of the samples planned to be used
as reference.}

\item{outfile.prefix}{the prefix of the name of the output file. The suffix '.bgz' will
be appended to this name prefix after compression.}

\item{out.pdf}{the name of the output pdf file.}

\item{appendIndex.outfile}{if TRUE (default), the results will be appended regularly on
the output file which will be ultimately indexed. This is recommend when a large number
of bins are analyzed. If FALSE, a data.frame with the bin counts will be returned and no
file are created.}

\item{chunk.size}{the number of bins to analyze at a time (for memory optimization).
Default is 100 000. Reduce this number if memory problems arise.}

\item{col.bc}{the column from 'files.df' defining the bin count file names.}
}
\value{
a list with
\item{bc}{the name of the file with the joined bin counts OR a data.frame with
these bin counts.}
\item{dstat}{a data.frame with for each sample its D-statistic (average correlation
with other reference samples).}
\item{pc.1.3}{a matrix with the first 3 principal components for all samples.}
\item{cor.pw}{a matrix with correlation for any pair of samples.}
}
\description{
Join bin counts of the samples to analyze and compute some QC metrics to help
defining the set of reference samples/
}
\author{
Jean Monlong
}

