% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qc.samples.summary.R
\name{qc.samples.summary}
\alias{qc.samples.summary}
\title{QC sample summary}
\usage{
qc.samples.summary(qc.res)
}
\arguments{
\item{qc.res}{a list D statistic, PCA and correlation information, produced
by 'qc.samples'.}
}
\value{
a shiny app in the web browser
}
\description{
Interactive representation of the QC metrics on the reference samples.
}
\author{
Jean Monlong
}

