% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.bedix}
\alias{read.bedix}
\title{Retrieve a subset of an indexed BED file.}
\usage{
read.bedix(file, subset, col.names = NULL, as.is = TRUE)
}
\arguments{
\item{file}{the name of the file which the data are to be read from.
Should be compressed with |code{bgzip} and indexed by tabix algorithm.}

\item{subset}{a data.frame or GRanges object with the regions to subset from.
If a data.frame, it must have columns named 'chr', 'start' and 'end'.}

\item{col.names}{The header for the output object. Optional.}

\item{as.is}{controls the conversion of columns. Default is TRUE, i.e. columns are
converted into 'character', 'numeric', 'integer', etc. If FALSE, some columns might
be converted into factors (e.g. 'character' columns).}
}
\value{
a data.frame with the retrieved BED information.
}
\description{
Using \code{Rsamtools} functions to manipulate tabix file, the indexed BED file is read
and only regions defined by the desired subset are retrieved.
}
\author{
Jean Monlong
}

