% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sv.summary.interactive}
\alias{sv.summary.interactive}
\title{Interactive summary}
\usage{
sv.summary.interactive(res.df, merge.cons.bin = TRUE, height = "500px")
}
\arguments{
\item{res.df}{the data.frame with the results.}

\item{merge.cons.bin}{TRUE if the single-bin calls has been merged
(Default, 'merge' option in 'call.abnormal.cov'). FALSE for single-bin calls.}

\item{height}{the height of the plot in the webpage. Default is "500px".}
}
\value{
a shiny app in the web browser.
}
\description{
Interactive summary of the individual bins with abnormal coverage as detected per 'call.abnomal.cov'.
}
\author{
Jean Monlong
}

