% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sv.summary.interactive.R
\name{sv.summary.interactive}
\alias{sv.summary.interactive}
\title{Interactive summary}
\usage{
sv.summary.interactive(res.df, height = "500px")
}
\arguments{
\item{res.df}{the data.frame with the results.}

\item{height}{the height of the plot in the webpage. Default is "500px".}
}
\value{
a shiny app in the web browser.
}
\description{
Interactive summary of the individual bins with abnormal coverage as detected per 'call.abnomal.cov'.
}
\author{
Jean Monlong
}

