% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sv.summary.R
\name{sv.summary}
\alias{sv.summary}
\title{Abnormal regions summary}
\usage{
sv.summary(res.df, out.pdf = NULL, print = TRUE)
}
\arguments{
\item{res.df}{the data.frame with the results.}

\item{out.pdf}{the name for the output PDF file to create. If NULL (default), the graphs are either displayed directly or returned as a list.}

\item{print}{Should the graphs be displayed ? If 'out.pdf' though, this parameter is not used as no graph will be displayed.}
}
\value{
A list with :
\item{graphs.l}{a list of ggplot graphs}
\item{info.df}{a data.frame with different summary statistics}
}
\description{
Summary of regions with abnormal coverage as detected by 'call.abnomal.cov'. It's a static version of 'sv.summary.interactive'. It produces graphs to to get an idea of the quality of the calls. For normal samples we would expect: the amount of called genome in each sample to be somewhat similar; the copy number estimates to locate near integer values; No systematic calls (i.e. called in all samples). The output graphs represent the amount of affected genome across samples, the copy number estimates distribution, the frequency distribution.
}
\details{
A list is returned. Its first element is a list of ggplot object with the different graphs. The second is a data.frame with summary information used to create the graphs.
}
\author{
Jean Monlong
}

