% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tn.norm.qc.R
\name{tn.norm.qc}
\alias{tn.norm.qc}
\title{QC graphs for Targeted Normalization}
\usage{
tn.norm.qc(norm.stats, out.pdf = "normStats-QC.pdf", bin.size = FALSE)
}
\arguments{
\item{norm.stats}{the name of the file with the normalization statistics ('norm.stats' in 'tn.norm' function) or directly a 'norm.stats' data.frame.}

\item{out.pdf}{the name of the PDF file to create.}

\item{bin.size}{Should the metrics be compared to the bin size. Default is FALSE. Useful if the bins are not of equal size.}
}
\value{
the name of the created PDF file.
}
\description{
Following targeted normalization, some Quality Control graphs are produced to see if some bins were problematic. Several graphs will be created. First the distribution of each normalization statistic: correlation distance with the last supporting bin; average normalized coverage; number of outlier samples removed. Then pairwise comparison of these metrics are displayed.
}
\author{
Jean Monlong
}

