% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tn.norm}
\alias{tn.norm}
\title{Targeted-normalization of bin counts}
\usage{
tn.norm(bc, cont.sample, ref.samples, nb.support.bins = 1000, bins = NULL,
  save.support.bins = TRUE, z.poisson = FALSE, aberrant.cases = TRUE)
}
\arguments{
\item{bc}{a matrix or data.frame with the bin counts (bin x sample).}

\item{cont.sample}{the sample to use as baseline for the pairwise normalization.
All the samples will be normalized to it.}

\item{ref.samples}{a vector with the names of the samples to be used as reference.}

\item{nb.support.bins}{the number of bins to use for the normalization.}

\item{bins}{a vector the names of the bins to normalize. If NULL (default), all
bins are normalized.}

\item{save.support.bins}{if TRUE (default) the bins used for the normalization are
saved in the output object 'norm.stats'.}

\item{z.poisson}{Should the Z-score be computed as an normal-poisson hybrid (see
Details). Default is FALSE.}

\item{aberrant.cases}{if TRUE (default) a more robust (but sligthly longer) normalization
is performed on cases to deal with potential large chromosomal aberrations. In practice,
it is recommended for cancer but can be turned off if less than 20% of the genome is expected
to be affected.}
}
\value{
a list with
\item{norm.stats}{a data.frame witht some metrics about the normalization of each
bin (row) : correlation with worst supporting bin; coverage average and standard
deviation; number of outlier reference samples; supporting bins.}
\item{bc.norm}{a matrix with the normalized bin counts (bin x sample).}
\item{z}{a matrix with the Z-scores for each bin and sample (bin x sample).}
\item{fc}{a matrix with the copy-number coefficients for each bin and sample (
bin x sample).}
\item{nb.support.bins, cont.sample, z.poisson}{a backup of the input parameters.}
}
\description{
Bin counts are normalized one bin at a time, using a subset of the bins that look
similar across the reference samples.
}
\details{
The Z-score is computed by substracting the bin count by the average bin count
across the reference samples and dividing by their standard deviation. If
'z.poisson' is TRUE, a score using Poisson distribution is also computed, using
the average bin count as an estimator of the lambda. Then the score with the lowest
absolute value is kept. This hybrid Z-score is to be used when some regions have low
coverage where it is more robust to use Poisson assumptions.
}
\author{
Jean Monlong
}

