% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tn.norm.R
\name{tn.norm}
\alias{tn.norm}
\title{Targeted-normalization of bin counts}
\usage{
tn.norm(bc, cont.sample, nb.support.bins = 1000, bins = NULL,
  save.support.bins = TRUE, norm = c("1pass", "bootstrap"))
}
\arguments{
\item{bc}{a matrix or data.frame with the bin counts (bin x sample).}

\item{cont.sample}{the sample to use as baseline for the pairwise normalization.
All the samples will be normalized to it.}

\item{nb.support.bins}{the number of bins to use for the normalization.}

\item{bins}{a vector the names of the bins to normalize. If NULL (default), all
bins are normalized.}

\item{save.support.bins}{if TRUE (default) the bins used for the normalization are
saved in the output object 'norm.stats'.}

\item{norm}{the type of normalization. '1pass' (default) means one pass of normalization. Other options is 'bootstrap'.}
}
\value{
a list with
\item{norm.stats}{a data.frame witht some metrics about the normalization of each
bin (row) : correlation with worst supporting bin ('d.max'); coverage average ('m') and standard deviation ('sd'); number of outlier reference samples ('nb.remove'); supporting bins.}
\item{bc.norm}{a matrix with the normalized bin counts (bin x sample).}
\item{nb.support.bins, cont.sample}{a backup of the input parameters.}
}
\description{
Bin counts are normalized one bin at a time, using a subset of the bins that look
similar across the reference samples.
}
\details{
A specific set of bins, defined by 'bins=', can de normalized using all the bins in 'bc'. The bin names in 'bins=' should be the chr and start position as in '1-501' (for chr 1, start 501).

The default approach ('norm="1pass"') looks for supporting across all samples. A more robust approach, finds supporting bins on different subset of samples with a bootstrap appoach. 'norm="bootstrap"' normalization is better and recommended for small bins, but it takes more time to run.
}
\author{
Jean Monlong
}

