% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/weight.ref.pca.R
\name{weight.ref.pca}
\alias{weight.ref.pca}
\title{Weight sample difference from a set of reference samples}
\usage{
weight.ref.pca(pca.mat, ref.samples, plot = FALSE, output.dist = FALSE)
}
\arguments{
\item{pca.mat}{a matrix with the principal components as columns and the sample names as row names. The first two components are used.}

\item{ref.samples}{the names of the samples used as reference.}

\item{plot}{should some graphs be displayed. Default is FALSE.}

\item{output.dist}{should the distance to the reference sample centroid be outputed instead of the weights. Default is FALSE.}
}
\value{
a vector with the weights for each sample.
}
\description{
Compute weight that represent how similar samples are to the reference samples, from the PCA. These weights are used to adjust the stringency of the threshold for the calls. Indeed, some subtle 'abnormal' signal detected might be due to fact that a sample is globally different from the reference samples, hence a more stringent threshold is advised.
}
\author{
Jean Monlong
}

