% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/z.comp.R
\name{z.comp}
\alias{z.comp}
\title{Z-score computation}
\usage{
z.comp(bc.f, files.df, ref.samples = NULL, z.poisson = FALSE,
  nb.cores = 1, chunk.size = NULL, out.msd.f = "ref-msd.tsv",
  append = FALSE)
}
\arguments{
\item{bc.f}{the path to the normalized bin count file.}

\item{files.df}{a data.frame with the file paths.}

\item{ref.samples}{a vector with the samples to use as reference. If 'NULL' (default) all samples in the bin count file are used as reference.}

\item{z.poisson}{Should the Z-score be computed as an normal-poisson hybrid (see
Details). Default is FALSE.}

\item{nb.cores}{the number of cores to use.}

\item{chunk.size}{the chunk size. If NULL (Default), no chunks are used.}

\item{out.msd.f}{the name of the file to write the mean/sd information. If 'NULL' no file is created.}

\item{append}{should the Z-scores be appended to existing files. Default is FALSE.}
}
\value{
a list with
\item{z}{a data.frame with the Z-scores for each bin and sample (bin x sample).}
\item{fc}{a data.frame with the fold-change compared to the average bin count in
the reference samples for each bin and sample (bin x sample).}
\item{msd}{the mean, standard deviation and number of removed outlier samples in each bin.}
}
\description{
Z-score computation from bin count and/or mean/sd metrics on the reference samples
}
\author{
Jean Monlong
}

