% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.comp.R
\name{z.comp}
\alias{z.comp}
\title{Z-score computation}
\usage{
z.comp(bc.f, norm.stats.f, files.df, z.poisson = FALSE, nb.cores = 1,
  chunk.size = 10000, append = FALSE)
}
\arguments{
\item{bc.f}{the path to the normalized bin count file.}

\item{norm.stats.f}{the name of the file with the statistic of the targeted normalization run.}

\item{files.df}{a data.frame with the file paths.}

\item{z.poisson}{Should the Z-score be computed as an normal-poisson hybrid (see
Details). Default is FALSE.}

\item{nb.cores}{the number of cores to use.}

\item{chunk.size}{the chunk size. Default is 1e4. If NULL, no chunks are used.}

\item{append}{should the Z-scores be appended to existing files. Default is FALSE.}
}
\value{
a list with
\item{samples}{a vector with the names of the samples analyzed.}
\item{z.poisson}{was Normal-Poisson hybrid Z-score score computed.}
}
\description{
Z-score computation from bin count and/or mean/sd metrics on the reference samples.
}
\details{
The Z-score is computed by substracting the bin count by the average bin count
across the reference samples and dividing by their standard deviation. If
'z.poisson' is TRUE, a score using Poisson distribution is also computed, using
the average bin count as an estimator of the lambda. Then the score with the lowest
absolute value is kept. This hybrid Z-score is to be used when some regions have low
coverage where it is more robust to use Poisson assumptions.
}
\author{
Jean Monlong
}

