% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/z.thres.cons.bins.R
\name{z.thres.cons.bins}
\alias{z.thres.cons.bins}
\title{Z-score thresholding using bin consecutiveness.}
\usage{
z.thres.cons.bins(z.df, plot = FALSE, pvalues = FALSE)
}
\arguments{
\item{z.df}{a data.frame with at least 'chr', 'start', 'end' and 'z' columns.}

\item{plot}{should some graphs be displayed. Default if FALSE.}

\item{pvalues}{should the P-values be deduced from these thresholds. Default is FALSE. Note: the P-values and Q-values derive from custom 'recipes' and should be taken as informative only, for further filtering or priorization.}
}
\value{
a list with:
\item{dup.th}{the threshold for positive Z-scores (duplication signal)}
\item{del.th}{the threshold for negative Z-scores (deletion signal)}
\item{nb.ab.bins}{the number of abnormal bins}
\item{prop.ab.bins}{the proportion of abnormal bins}
\item{z.df}{a data.frame, similar to the input 'z.df' with new columns: 'abnormal' for bins with abnormal Z-scores and potentially 'pv'/'qv' for P-values/Q-values.}
}
\description{
Find a threshold on the Z-scores associated with how many consecutive bins are found.
}
\author{
Jean Monlong
}
\keyword{internal}

